I think the minQueue is working properly now.
The only problem is with a weird memory leak... I am calling free but it isn't effective for some reason. Weird. I am not going to worry about it as there is not real cleanup needed in the actual operating system implementation as
this is actually below that level. Still annoying :P. Now to get the actual memory allocater to work.