Memory Allocater seems to work for now!
I haven't done extensive testing and should add more tests later but this is enough for now. I need to do some cleaning up, including adding custom namespaces to some of the functions and making sure all comments match and are legible.
If you want to run the memalloctest, make sure to run with -fno-stack-protector because who cares lol (probably a problem I should deal with but nah).